Data for the related analysis.
This dataset contains data about the physical and chemical properties of the Li-ion silicate cathodes. These properties can be useful to predict the class of a Li-ion battery. These batteries can be classified on the basis of their crystal system. Three major classes of crystal system include: monoclinic, orthorhombic and triclinic.